Fix Select error

+++ index.php
@@ -40,26 +40,30 @@ $result = mysqli_query($mysqli, "SELECT * FROM cars ORDER BY id DESC"); // using
         <td>Update</td>
     </tr>
     <?php 
+if($result !== false) {
     while($res = mysqli_fetch_array($result)) { 		
-        echo "<tr>";
-        echo "<td>".$res['license_number']."</td>";
-        echo "<td>".$res['brand']."</td>";
-        echo "<td>".$res['model']."</td>";
-        echo "<td>".$res['car_plate']."</td>";
-        echo "<td>".$res['km']."</td>";
-        echo "<td>".$res['category']."</td>";
-        echo "<td>".$res['type']."</td>";
-        echo "<td>".$res['comments']."</td>";
-        echo "<td>".$res['discharge_date']."</td>";
-        echo "<td>".$res['color']."</td>";
-        echo "<td>".$res['extras']."</td>";
-        echo "<td>".$res['car_image']."</td>";
-        echo "<td>".$res['price']."</td>";
-        echo "<td>".$res['doors']."</td>";
-        echo "<td>".$res['city']."</td>";
-        echo "<td>".$res['lat']."</td>";
-        echo "<td>".$res['lng']."</td>";
-        echo "<td><a href=\"edit.php?id=$res[id]\">Edit</a> | <a href=\"delete.php?id=$res[id]\" onClick=\"return confirm('Are you sure you want to delete?')\">Delete</a></td>";		
+            echo "<tr>";
+            echo "<td>".$res['license_number']."</td>";
+            echo "<td>".$res['brand']."</td>";
+            echo "<td>".$res['model']."</td>";
+            echo "<td>".$res['car_plate']."</td>";
+            echo "<td>".$res['km']."</td>";
+            echo "<td>".$res['category']."</td>";
+            echo "<td>".$res['type']."</td>";
+            echo "<td>".$res['comments']."</td>";
+            echo "<td>".$res['discharge_date']."</td>";
+            echo "<td>".$res['color']."</td>";
+            echo "<td>".$res['extras']."</td>";
+            echo "<td>".$res['car_image']."</td>";
+            echo "<td>".$res['price']."</td>";
+            echo "<td>".$res['doors']."</td>";
+            echo "<td>".$res['city']."</td>";
+            echo "<td>".$res['lat']."</td>";
+            echo "<td>".$res['lng']."</td>";
+            echo "<td><a href=\"edit.php?id=$res[id]\">Edit</a> | <a href=\"delete.php?id=$res[id]\" onClick=\"return confirm('Are you sure you want to delete?')\">Delete</a></td>";		
+        }
+    } else {
+        echo "Error en la consulta: Parece que no hay datos en la tabla.";
     }
     include 'debug.php';
     $data = 'Eso es el console_log';
